// src/modules/province/types/province.type.ts
export type TProvince = {
  id: string;
  name: string;
  code: number;
  division_type: string;
  codename: string;
  phone_code: number;
};

export type TDistrict = {
  id: string;
  name: string;
  code: number;
  codename: string;
  division_type: string;
  short_codename: string;
};

export type TWard = {
  id: string;
  name: string;
  code: number;
  codename: string;
  division_type: string;
  short_codename: string;
};